Create style3.css
Broadway
In this project, you will use properties such as display and position to improve the layout of the landing page for a fictional design firm, Broadway Design.

The site has some style rules to begin with. You will improve the layout and positioning of the navigation menu at the top of the page and the three supporting sections (Design, Develop, Deploy) below the image.

If you get stuck during this project or would like to see an experienced developer work through it, click “Get Unstuck“ to see a project walkthrough video.